feat(fixtures):generattion of fixture 'serde' code
We moved all unit-tests into integration tests and automated the
generation of serialization/deserialization code required for
our test-datastructures. That way, we run on stable from the get-go.

Additinally, a Makefile was added to keep track of the processes
needed to get the data we want. It also is the foundation for what's
going to come in terms of using existing YAML implementations to compare
with our one.

That way, we are not depending on certain issues in syntex anymore,
and the entire system looks way cleaner.