tests(infrastructure): add examples 2.1 to 2.12
Each example is serialized by the python implementation just to
see if it can produce the desired output, in conjunction with a
Rust type that is serialized in a per-example test to check against
the string produce by the python implementation.

We should be able to derive the required set of options options to
configure the output accordingly.

We are missing about 12 additional examples, and should probably
start filling in the options we see so far.

[skip ci]