refactor(ser): remove all [inline] statements
That way, the optimizer may decide when to inline, and when not.
When looking at the standard library, you will see that most of the
small methods are not inline (anymore) as it put compile times to a
crawl and didn't necessarily improve performance as too much code
was produced in the process.

[skip ci]