feat(ser): canonical presentation
Added presentation details to allow for the canonical presentation mode,
which nose has a custom constructor to ease generating the correct
combination of options.

Added tests for it, as well as for 'explicit mapping entry' which allows
for complex keys, and is also part of the requirements for canonical
presentation.

The only question left is how to specify newlines, and who takes care
of it accordingly.

[skip ci]