refactor(ser): `PresentationDetails` on Serializer
`PresentationDetails` still carry all options, but options related
to whitespace handling and encoding are now part of the
`FormatterOptions`, which are the only options the internal Formatter
should need.

This fixes a few ownership issues, and prepares the complete removal
of the StandardFormatter to ease the signature of the upcoming
`encode_scalar`.

`encode_scalar` is currently a function, that requires implementation.
The latter will be quite complex, and thus needs to be done in plenty
of steps/sessions.

[skip ci]