refactor(ser): merge Formatter into Serializer
This simplifies method calls and the overall architecture, and should
help us getting a working serializer implementation.

The formatter could be seen as performance improvement, as it would
allow to handle whitespace very efficiently.

However, we are not at a spot yet where we could put in complexitiy
to possibly gain performance.

[skip ci]