tests(ser): block mapping tests
A test for how to handle null value hiding within flow-sequences led
to more complex scenarios that involved mappings, which now need to
be implemented it seems ;).

The good thing is that we should end up with somewhat final data
structures for our internal state to handle the flow-style as well
as null-value hiding correctly.

Makefile adjusted to run unit-tests only (as these are the only ones
with a chance to work).

Added two new structures to help test simple mappings, with one
and two optional keys respectively.

[skip ci]