feat(ser): JSON and YAML escaping - WIP
We added plenty of cases to verify that json and yaml compatible
encoding works.

However, we still might not deal with all JSON characters correctly,
nor do we handle YAML whitespace. For instance, if there is a whitespace
suffix or whitespace prefix, we need to enforce single quotes at least !

[skip ci]