feat(ser): handle control character escaping
Escape control characters as identified by unicode. This covers just a
few remaining control characters, both for yaml and json.

Everything else is considered printable, which is nice for things like
smileys, chinese and russian, which totally show up in your editor
as they are ... . Probably works for arabic as well.

[skip ci]