feat(ser): simple explicit entries handling
It's simple because it just deals with adding newlines before colons
in block mode and question marks in front of keys. This can all be done
without peeking the value, which might actually require explicit mode
even if we didn't say so in the options.

The latter is also depending on the desired structure style ... .

[skip ci]